% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{A shiny app to import QFeatures objects.}
\usage{
import(colData = NULL, assayData = NULL, maxSize = 1000)
}
\arguments{
\item{colData}{A data frame that contains the sample table.}

\item{assayData}{A data frame that contains the input table.}

\item{maxSize}{An integer that changes the shiny.maxRequestSize value, in MB.}
}
\value{
The "import" Shiny app object.
}
\description{
import is a simple graphical interface to import bulk and single-cell proteomics data.
The app uses the \code{\link[QFeatures]{readQFeatures}} function from the QFeatures package to convert simple tables (single or multiple, CSV or TSV) to a QFeatures object.
The app allows users to convert tables to a QFeatures object.
}
\examples{
library(QFeaturesGUI)

data("sampleTable")
data("inputTable")
app <- import(colData = sampleTable, assayData = inputTable, maxSize = 100)

if (interactive()) {
    shiny::runApp(app)
}

}
