% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processQFeatures_WIP.R
\name{processQFeatures_WIP}
\alias{processQFeatures_WIP}
\title{A shiny app to process QFeatures objects.}
\usage{
processQFeatures_WIP(colData = NULL, assayData = NULL)
}
\arguments{
\item{colData}{A dataframe that contains the sample table.}

\item{assayData}{A dataframe that contains the input table.}
}
\value{
Return the "importQFeatures" shiny app object.
}
\description{
processQFeatures is a simple graphical interface to process QFeatures object.
The app is divided into mutliple sections:
\itemize{
\item The first section (Import) of the app allow to convert tables to a QFeatures object.
\item The second section (Pre-processing) of the app allow to perform some optional pre-processing steps.
}
}
\examples{
library(QFeaturesGUI)

data("sampleTable")
data("inputTable")
app <- importQFeatures(colData = sampleTable, assayData = inputTable)

if (interactive()) {
    shiny::runApp(app)
}

}
