% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scpGUI.R
\name{scpGUIImport}
\alias{scpGUIImport}
\title{A simple graphical interface to import SCP data}
\usage{
scpGUIImport()
}
\value{
Launch the scpGUIImport shiny app.
}
\description{
scpGUI is a simple graphical interface to handle SCP data.
The first page of the app allow to convert two tables (Input and Sample) to a QFeatures object.
This process is performed with the \code{\link[scp]{readSCP}} function from the scp package.
}
\examples{
scpGUIImport() #start a shiny app

}
