% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise}
\alias{visualise}
\title{Launch a Shiny application to visualise QFeatures objects}
\usage{
visualise(qfeatures = NULL, maxSize = 100)
}
\arguments{
\item{qfeatures}{Optional \linkS4class{QFeatures} object to visualise,
or a character string specifying the path to an \code{.rds} file
containing one. If omitted or \code{NULL}, the app displays a startup
modal for uploading a file or loading the bundled demo.}

\item{maxSize}{An integer that changes the \code{shiny.maxRequestSize}
value, in MB. This controls the maximum upload size for the startup
\code{.rds} file upload modal.}
}
\value{
The visualise Shiny application.
}
\description{
\code{visualise()} launches an interactive Shiny application
that allows users to visualise a \linkS4class{QFeatures} object.

The input \code{qfeatures} can be provided as an in-memory
\linkS4class{QFeatures} object, as a path to an \code{.rds} file
containing one, or omitted. If omitted, the application prompts the user
to upload a \linkS4class{QFeatures} object from an \code{.rds} file
or use the bundled demo dataset.
}
\examples{

library(QFeaturesGUI)


app <- visualise()

if (interactive()) {
    shiny::runApp(app)
}
}
