% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importQFeatures.R
\name{importQFeatures}
\alias{importQFeatures}
\title{A shiny app to import QFeatures objects.}
\usage{
importQFeatures(colData = NULL, assayData = NULL, maxSize = 1000)
}
\arguments{
\item{colData}{A dataframe that contains the sample table.}

\item{assayData}{A dataframe that contains the input table.}

\item{maxSize}{An integer that change shiny.maxRequestSize value, this value has to be in Mb.}
}
\value{
Return the "importQFeatures" shiny app object.
}
\description{
importQFeatures is a simple graphical interface to import bulk and single-cell proteomic data.
The app use the \code{\link[QFeatures]{readQFeatures}} function from the QFeatures package to convert simple table (single or multiple, csv or tsv) to a QFeatures object.
The app allow to convert tables to a QFeatures object.
}
\examples{
library(QFeaturesGUI)

data("sampleTable")
data("inputTable")
app <- importQFeatures(colData = sampleTable, assayData = inputTable, maxSize = 100)

if (interactive()) {
    shiny::runApp(app)
}

}
