% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processQFeatures.R
\name{processQFeatures}
\alias{processQFeatures}
\title{A shiny app to process QFeatures objects.}
\usage{
processQFeatures(qfeatures, initialSets = seq_along(qfeatures))
}
\arguments{
\item{qfeatures}{a \code{QFeatures} object that will be process by the application}
}
\value{
Return the "processQFeatures" shiny app object.
}
\description{
processQFeatures is a simple graphical interface to process QFeatures object.
}
\examples{

library(QFeatures)
library(QFeaturesGUI)

data("sampleTable")
data("inputTable")
qfeatures <- readQFeatures(inputTable,
    colData = sampleTable,
    runCol = "Raw.file")
app <- importQFeatures(qfeatures, initialSets = seq_along(qfeatures))

if (interactive()) {
    shiny::runApp(app)
}

}
