% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scpGUI.R
\name{scpGUI}
\alias{scpGUI}
\title{A simple graphical interface to import SCP data}
\usage{
scpGUI()
}
\value{
Return the "scpGUI" shiny app object.
}
\description{
scpGUI is a simple graphical interface to handle SCP data.
The first page of the app allow to convert two tables (Input and Sample) to a QFeatures object.
This process is performed with the \code{\link[scp]{readSCP}} function from the scp package.
}
\examples{
library(scpGUI)

app <- scpGUI()

if (interactive()) {
    shiny::runApp(app, ...)
}

}
