% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeQFeatures.R
\name{visualizeQFeatures}
\alias{visualizeQFeatures}
\title{Launch a Shiny application to visualize QFeatures objects}
\usage{
visualizeQFeatures(qfeatures = NULL, maxSize = 100)
}
\arguments{
\item{maxSize}{An integer that changes the \code{shiny.maxRequestSize}
value, in MB. This controls the maximum upload size for the startup
\code{.rds} file upload modal.}
}
\description{
\code{processQFeatures()} launches an interactive Shiny application
that allows users to visualize a \linkS4class{QFeatures} object.

The input \code{qfeatures} can be provided as an in-memory
\linkS4class{QFeatures} object, as a path to an \code{.rds} file
containing one, or omitted. If omitted, the application prompts the user
to upload a \linkS4class{QFeatures} object from an \code{.rds} file.
}
\examples{

library(QFeaturesGUI)


app <- visualizeQFeatures()

if (interactive()) {
    shiny::runApp(app)
}
}
